refactor: 音声ライブラリ API Router へのモジュール化
