+++ run.py
@@ -1,5 +1,4 @@
 import argparse
-import asyncio
 import json
 import multiprocessing
 import os
@@ -8,16 +7,14 @@ import sys
 from collections.abc import AsyncIterator, Awaitable, Callable
 from contextlib import asynccontextmanager
 from functools import lru_cache
-from io import BytesIO, TextIOWrapper
+from io import TextIOWrapper
 from pathlib import Path
 from tempfile import NamedTemporaryFile
 from typing import Annotated, Any, Optional
 
 import soundfile
 import uvicorn
-from fastapi import Depends, FastAPI, HTTPException
-from fastapi import Path as FAPath
-from fastapi import Query, Request, Response
+from fastapi import FastAPI, HTTPException, Query, Request, Response
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.openapi.utils import get_openapi
 from fastapi.responses import JSONResponse
@@ -27,11 +24,9 @@ from starlette.middleware.errors import ServerErrorMiddleware
 from starlette.responses import FileResponse
 
 from voicevox_engine import __version__
-from voicevox_engine.app.dependencies import (
-    check_disabled_mutable_api,
-    deprecated_mutable_api,
-)
+from voicevox_engine.app.dependencies import deprecated_mutable_api
 from voicevox_engine.app.routers import (
+    library,
     preset,
     setting,
     speaker,
@@ -49,8 +44,6 @@ from voicevox_engine.metas.MetasStore import MetasStore, construct_lookup
 from voicevox_engine.model import (
     AudioQuery,
     BaseLibraryInfo,
-    DownloadableLibraryInfo,
-    InstalledLibraryInfo,
     MorphableTargetInfo,
     StyleIdNotFoundError,
     SupportedDevicesInfo,
@@ -371,82 +364,9 @@ def generate_app(
     app.include_router(speaker.generate_router(get_core, metas_store, root_dir))
 
     if engine_manifest_data.supported_features.manage_library:
-
-        @app.get(
-            "/downloadable_libraries",
-            response_model=list[DownloadableLibraryInfo],
-            response_description="ダウンロード可能な音声ライブラリの情報リスト",
-            tags=["音声ライブラリ管理"],
+        app.include_router(
+            library.generate_router(engine_manifest_data, library_manager)
         )
-        def downloadable_libraries() -> list[DownloadableLibraryInfo]:
-            """
-            ダウンロード可能な音声ライブラリの情報を返します。
-            """
-            if not engine_manifest_data.supported_features.manage_library:
-                raise HTTPException(
-                    status_code=404, detail="この機能は実装されていません"
-                )
-            return library_manager.downloadable_libraries()
-
-        @app.get(
-            "/installed_libraries",
-            response_model=dict[str, InstalledLibraryInfo],
-            response_description="インストールした音声ライブラリの情報",
-            tags=["音声ライブラリ管理"],
-        )
-        def installed_libraries() -> dict[str, InstalledLibraryInfo]:
-            """
-            インストールした音声ライブラリの情報を返します。
-            """
-            if not engine_manifest_data.supported_features.manage_library:
-                raise HTTPException(
-                    status_code=404, detail="この機能は実装されていません"
-                )
-            return library_manager.installed_libraries()
-
-        @app.post(
-            "/install_library/{library_uuid}",
-            status_code=204,
-            tags=["音声ライブラリ管理"],
-            dependencies=[Depends(check_disabled_mutable_api)],
-        )
-        async def install_library(
-            library_uuid: Annotated[str, FAPath(description="音声ライブラリのID")],
-            request: Request,
-        ) -> Response:
-            """
-            音声ライブラリをインストールします。
-            音声ライブラリのZIPファイルをリクエストボディとして送信してください。
-            """
-            if not engine_manifest_data.supported_features.manage_library:
-                raise HTTPException(
-                    status_code=404, detail="この機能は実装されていません"
-                )
-            archive = BytesIO(await request.body())
-            loop = asyncio.get_event_loop()
-            await loop.run_in_executor(
-                None, library_manager.install_library, library_uuid, archive
-            )
-            return Response(status_code=204)
-
-        @app.post(
-            "/uninstall_library/{library_uuid}",
-            status_code=204,
-            tags=["音声ライブラリ管理"],
-            dependencies=[Depends(check_disabled_mutable_api)],
-        )
-        def uninstall_library(
-            library_uuid: Annotated[str, FAPath(description="音声ライブラリのID")]
-        ) -> Response:
-            """
-            音声ライブラリをアンインストールします。
-            """
-            if not engine_manifest_data.supported_features.manage_library:
-                raise HTTPException(
-                    status_code=404, detail="この機能は実装されていません"
-                )
-            library_manager.uninstall_library(library_uuid)
-            return Response(status_code=204)
 
     @app.post("/initialize_speaker", status_code=204, tags=["その他"])
     def initialize_speaker(
